requirements.txt
absl-py==2.1.0
aider-install==0.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
audioop-lts==0.2.1
auto-py-to-exe==2.46.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
branca==0.8.1
cachetools==5.5.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.9
contourpy==1.3.1
cors==1.0.1
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
django-countries==7.6.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
Eel==0.18.2
english-words==2.0.1
et_xmlfile==2.0.0
executing==2.1.0
fairlearn==0.12.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpeg_downloader==0.4.0
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.2
folium==0.20.0
fonttools==4.55.3
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gcloud==0.18.3
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.8.6
googleapis-common-protos==1.70.0
gradio==5.23.3
gradio_client==1.8.0
greenlet==3.2.3
groovy==0.1.2
groq==1.0.0
grpcio==1.75.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==25.1.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwcrypto==1.5.6
keyboard==0.13.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maskpass==0.3.7
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml_collections==1.0.0
mouse==0.7.1
MouseInfo==0.1.3
multidict==6.2.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
narwhals==1.22.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
oauth2client==4.1.3
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
panda==0.3.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
playsound==1.2.2
plotly==6.1.2
plyer==2.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyjokes==0.8.3
PyJWT==2.10.1
pymongo==4.13.2
PyMsgBox==1.0.9
PyMuPDF==1.27.1
pynput==1.8.1
pyparsing==3.2.1
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
Pyrebase4==4.8.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==3.2.1
python-jwt==4.1.0
python-multipart==0.0.20
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-file==3.0.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
ruff==0.11.4
safehttpx==0.1.6
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.15.1
seaborn==0.13.2
selenium==4.38.0
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.3
speedtest-cli==2.1.3
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.1
streamlit==1.41.1
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.3.1
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
twilio==9.5.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
uv==0.8.22
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.11.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wordcloud==1.9.4
wsproto==1.2.0
xyzservices==2025.4.0
yarl==1.18.3
yt-dlp==2025.10.22
zope.event==5.1.1
zope.interface==7.2
